+++ app/src/main/java/com/boosters/promise/ui/promisesetting/PromiseSettingActivity.kt
@@ -1,4 +1,4 @@
-package com.boosters.promise.ui.promise
+package com.boosters.promise.ui.promisesetting
 
 import android.content.Intent
 import android.os.Build.VERSION
@@ -15,9 +15,9 @@ import com.boosters.promise.databinding.ActivityPromiseSettingBinding
 import com.boosters.promise.ui.invite.InviteActivity
 import com.boosters.promise.ui.invite.model.UserUiState
 import com.boosters.promise.ui.place.PlaceSearchDialogFragment
-import com.boosters.promise.ui.promise.adapter.PromiseMemberListAdapter
-import com.boosters.promise.ui.promise.model.PromiseSettingEvent
-import com.boosters.promise.ui.promise.model.PromiseSettingUiState
+import com.boosters.promise.ui.promisesetting.adapter.PromiseMemberListAdapter
+import com.boosters.promise.ui.promisesetting.model.PromiseSettingEvent
+import com.boosters.promise.ui.promisesetting.model.PromiseSettingUiState
 import com.google.android.material.datepicker.MaterialDatePicker
 import com.google.android.material.snackbar.Snackbar
 import com.google.android.material.timepicker.MaterialTimePicker
