+++ app/src/main/java/com/boosters/promise/ui/promisesetting/model/PromiseSettingEvent.kt
@@ -1,4 +1,4 @@
-package com.boosters.promise.ui.promise.model
+package com.boosters.promise.ui.promisesetting.model
 
 enum class PromiseSettingEvent {
     SELECT_DATE, SELECT_TIME, SELECT_PLACE, SELECT_MEMBER
