+++ app/src/main/res/layout/activity_promise_setting.xml
@@ -3,10 +3,10 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
-        <import type="com.boosters.promise.ui.promise.model.PromiseSettingEvent" />
+        <import type="com.boosters.promise.ui.promisesetting.model.PromiseSettingEvent" />
         <variable
             name="promiseSettingViewModel"
-            type="com.boosters.promise.ui.promise.PromiseSettingViewModel" />
+            type="com.boosters.promise.ui.promisesetting.PromiseSettingViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
