+++ app/src/main/res/layout/item_promise.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="promiseUiState"
-            type="com.boosters.promise.ui.promise.model.PromiseUiState" />
+            type="com.boosters.promise.ui.promisesetting.model.PromiseUiState" />
     </data>
 
     <com.google.android.material.card.MaterialCardView
